refactor : 약속 현황 화면의 블록을 클릭 시 나오는 바텀시트를 공통 컴포넌트 영역으로 이동
