+++ presentation/src/main/java/com/yapp/growth/presentation/component/PlanzBottomSheetContent.kt
@@ -13,12 +13,98 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import com.yapp.growth.domain.entity.TimeTable
 import com.yapp.growth.domain.entity.User
 import com.yapp.growth.presentation.R
+import com.yapp.growth.presentation.theme.Gray700
 import com.yapp.growth.presentation.theme.Gray800
 import com.yapp.growth.presentation.theme.MainPurple900
 import com.yapp.growth.presentation.theme.PlanzTypography
+import com.yapp.growth.presentation.util.getCurrentBlockDate
+import com.yapp.growth.presentation.util.toHour
+import com.yapp.growth.presentation.util.toPlanDate
+
+@Composable
+fun PlanzParticipantBottomSheetContent(
+    timeTable: TimeTable,
+    currentClickTimeIndex: Pair<Int, Int>,
+    currentClickUserData: List<User>,
+    onClickSelectPlan: (String) -> Unit = { },
+    isLeader: Boolean,
+) {
+    if (currentClickTimeIndex.first < 0 || currentClickTimeIndex.second < 0) return
+
+    val day = timeTable.availableDates[currentClickTimeIndex.first].split('T').first()
+    val hour = timeTable.minTime.toHour()
+    val time = "${day}T${hour}:00"
+
+    val respondUserText = StringBuilder()
+    currentClickUserData.forEachIndexed { index, user ->
+        when (index) {
+            0, 3, 7 -> respondUserText.append(user.userName)
+            2, 6 -> respondUserText.append(", ${user.userName.plus("\n")}")
+            else -> respondUserText.append(", ${user.userName}")
+        }
+    }
+
+    Column(
+        modifier = Modifier
+            .background(Color.White)
+            .padding(start = 20.dp, end = 20.dp)
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth(),
+            horizontalArrangement = Arrangement.SpaceBetween,
+        ) {
+            Text(
+                text = stringResource(R.string.planz_component_participant_bottom_sheet_content_promising_date),
+                style = PlanzTypography.subtitle2,
+                color = Gray700
+            )
+
+            Text(
+                text = time.getCurrentBlockDate(currentClickTimeIndex.second).toPlanDate(),
+                style = PlanzTypography.subtitle2,
+                color = MainPurple900
+            )
+        }
+
+        if (respondUserText.isNotEmpty()) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 12.dp),
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                Text(
+                    text = stringResource(R.string.planz_component_participant_bottom_sheet_content_respondent),
+                    style = PlanzTypography.subtitle2,
+                    color = Gray700
+                )
+
+                Text(
+                    text = respondUserText.toString(),
+                    style = PlanzTypography.caption,
+                    color = Gray800,
+                    textAlign = TextAlign.End
+                )
+            }
+        }
+
+        if (isLeader) {
+            PlanzBasicBottomButton(modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 20.dp),
+                text = stringResource(id = R.string.fix_plan_fix_button_title),
+                onClick = {
+                    onClickSelectPlan(time.getCurrentBlockDate(currentClickTimeIndex.second))
+                })
+        }
+    }
+}
 
 @Composable
 fun PlanzRespondentBottomSheetContent(
