+++ presentation/src/main/res/values/strings.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="planz_component_text_field_error_message_text">자를 초과하였습니다</string>
-    <string name="planz_component_respondent_bottom_sheet_content_title">응답자</string>
+    <string name="planz_component_participant_bottom_sheet_content_promising_date">약속 날짜</string>
+    <string name="planz_component_participant_bottom_sheet_content_respondent">응답자</string>
     <string name="planz_component_respondent_bottom_sheet_content_party_member">파티원</string>
     <string name="planz_component_alert_dialog_title">안내</string>
     <string name="planz_component_alert_dialog_positive_button_text">확인</string>
